requirements.txt
tensorflow==2.15.0
pillow==10.3.0
streamlit==1.35.0
keras==2.15.0